Working version for Roll Back 10-01-2025
This version has working image upload functionality but needs some work on the email harvesting logic. This is a good rollback version to start back up from. The R2 bucket logic is good and the the email logic is functional but needs work on the reply logic and the approval logic. This is the most advanced version so far.